requirements.txt
aiogram
fastapi
pycryptodome
pydantic
python-dotenv
uvicorn
pytest
httpx